Replace new parser to old parser
moved old parser to /Archived folder, for referencing in future.
